% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{f_cpp}
\alias{f_cpp}
\title{C++ implementation of the completeness fraction.}
\usage{
f_cpp(M, Lim, a = 1.5)
}
\arguments{
\item{M}{A numeric vector. The true GC magnitude}

\item{Lim}{Numeric with length 1. 50\% completeness fraction.}

\item{a}{Numeric with length 1. Rate at which the completeness fraction drops with M. Default to \code{a = 1.5}.}
}
\value{
The completeness fraction evaluated at the true magnitude \code{M}.
}
\description{
Compute the completeness fraction based on \deqn{f(M)=\frac{1}{\exp(\alpha(M- M_{\text{Lim}}))}.}
}
