\name{MATHPOP-package}
\alias{MATHPOP-package}
\alias{MATHPOP}
\docType{package}
\title{
\packageTitle{MATHPOP}
}
\description{
\packageDescription{MATHPOP}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{MATHPOP}
\packageIndices{MATHPOP}
For an overview and introduction of the package, see the \href{https://ddavidli.com/MATHPOP/}{MATHPOP webpage}. See \href{https://ddavidli.com/MATHPOP/vignette.html}{here} for a quick start on how to use the package.
}
\author{
\packageAuthor{MATHPOP}

Maintainer: \packageMaintainer{MATHPOP}
}
\references{
Li et al. (2024) "Discovery of Two Ultra-Diffuse Galaxies with Unusually Bright Globular Cluster Luminosity Functions via a Mark-Dependently Thinned Point Process (MATHPOP)", submitted to the Astrophysical Journal.
}
