% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meas_uncertain_mix_func.R
\name{meas_uncertain_mix_func}
\alias{meas_uncertain_mix_func}
\title{Fitting the mixture model with noisy point source data}
\usage{
meas_uncertain_mix_func(dat, n_iter, seed = 12345)
}
\arguments{
\item{dat}{A data frame that contains the point source data used to obtain the probabilistic GC catalog. Should contain the columns \code{C} for GC color; \code{F814W} for GC magnitudes; \code{C_err} for measurement uncertainty in color; \code{M_err} for measurement uncertainty in magnitudes.}

\item{n_iter}{An integer for number of iteration to jitter the sources with their measurement uncertainty.}

\item{seed}{An integer for the random seed. Default to 12345.}
}
\value{
A list with three objects \code{prob}, \code{par}, and \code{sim}. \code{prob} contains the computed probability that a source is a GC. \code{par} contains the inferred parameter values for the finite-mixture model. \code{sim} contains the jittered point source data.
}
\description{
Function to obtain probabilistic GC catalog based on the parametric finite-mixture model in Li et al. (2024).
}
