% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{phi_eM_cpp}
\alias{phi_eM_cpp}
\title{C++ implementation to calculate the density function of the noisy GCLF}
\usage{
phi_eM_cpp(M, mu, sigma, m0, b0, b1, n = 20L)
}
\arguments{
\item{M}{A numeric vector. The observed GC magnitude}

\item{mu}{Numeric with length 1. True GCLF TO point}

\item{sigma}{Numeric with length 1. True GCLF dispersion}

\item{m0}{Numeric with length 1. The offset magnitude}

\item{b0}{Numeric with length 1. The uncertainty when M = m0}

\item{b1}{Numeric with length 1. Rate at which the uncertainty increases with M}

\item{n}{Integer. Number of integration points used to marginalize out the true magnitude. Default to 20.}
}
\description{
C++ implementation to calculate the density function of the noisy GCLF
}
