% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psi_f.R
\name{psi_f}
\alias{psi_f}
\title{Function to compute the (unnormalized) truncated noisy GCLF}
\usage{
psi_f(M, mu, sigma, Lim, m0, b0, b1, a)
}
\arguments{
\item{M}{A numeric vector. Observed GC magnitudes.}

\item{mu}{A numeric value. True GCLF TO point.}

\item{sigma}{A numeric value. True GCLF dispersion.}

\item{Lim}{A numeric value. 50\% completeness limit.}

\item{m0}{A numeric value. Offset magnitude for measurement uncertainty.}

\item{b0}{A numeric value. Measurement uncertainty when \code{M = m0}.}

\item{b1}{A numeric value. Rate at which the measurement uncertainty increases as \code{M} increases.}

\item{a}{A numeric value. Rate at which the completeness fraction decreases as \code{M} increases.}
}
\value{
A numeric vector giving (unnormalized) truncated noisy GCLF evaluated at \code{M}.
}
\description{
Compute the (unnormalized) truncated noisy GCLF: \deqn{\varphi(m; \mu_{\text{TO}}, \sigma^2)f(m),}
where \eqn{\varphi(m; \mu_{\text{TO}}, \sigma^2)} is obtained from \code{\link[=phi_eM_cpp]{phi_eM_cpp()}}.
}
