% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sersic_ints.R
\name{Sersic_ints}
\alias{Sersic_ints}
\title{Function to compute the intensity at the locations of a point pattern \code{X} given a Sersic intensity profile}
\usage{
Sersic_ints(X, c, N, R_eff, e = 1, n = 0.5, theta = 0)
}
\arguments{
\item{X}{A two-column matrix of the GC point pattern locations. The first column is the x coordinate, the second column is the y coordinate.}

\item{c}{A vector of length two for the central location of a galaxy. The first entry is the x coordinate of the galactic center, the second entry is the y coordinate.}

\item{N}{A non-negative numeric value. Mean number of GC in a galaxy.}

\item{R_eff}{A non-negative numeric value. Half-number radius of the GC system of a galaxy.}

\item{e}{A non-negative numeric value. Aspect ratio of the GC system of a galaxy.}

\item{n}{A non-negative numeric value. Sersic index of the GC system of a galaxy.}

\item{theta}{A numeric value in \eqn{(0, 2\pi)}. Orientation angle (in radian) of the GC system of a galaxy.}
}
\value{
A numeric vector with the length equal to the number of row of \code{X} that gives the Sersic intensity at the locations specified by \code{X}.
}
\description{
Compute the intensity of a point pattern \code{X} based on the Sersic profile given by \deqn{\Lambda(\mathbf{X}; N, R_h, n) = \frac{N b_n^{2n}}{2\pi R_h^2 n \Gamma(2n)e}\exp\left(-b_n\left(\frac{r(\mathbf{X})}{R_h}\right)^{1/n}\right),}
where \deqn{r^2(\mathbf{X}) = ((\mathbf{X}_x - c_x)\cos(\vartheta) - (\mathbf{X}_y - c_y)\sin(\vartheta))^2 + ((\mathbf{X}_x- c_x)\sin(\vartheta)+ (\mathbf{X}_y-c_y)\cos(\vartheta))^2/e^2.}.
}
\keyword{Intensity}
\keyword{Sersic}
