% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_Y.R
\name{simulate_Y}
\alias{simulate_Y}
\title{Simulate GC locations and magnitudes (without measurement uncertainty) based on a Sersic profile and Gaussian GCLF.}
\usage{
simulate_Y(S, l0, c, N, R_eff, e, n, theta, mu, sigma)
}
\arguments{
\item{S}{A \code{SpatialPolygonsDataFrame} that gives the spatial domain on which the point pattern resides.}

\item{l0}{A positive numeric value. The intensity (/kpc^2) of GCs in the IGM.}

\item{c}{A matrix with two columns for the central locations of galaxies in \code{S}. The first column is the x coordinate of the galactic centers, the second column is the y coordinate.}

\item{N}{A (non-negative) numeric vector with length \code{nrow(c)}. Each element is the mean number of GCs in a galaxy.}

\item{R_eff}{A (non-negative) numeric vector with length \code{nrow(c)}. Each element is the half-number radius of the GC system of a galaxy.}

\item{e}{A (non-negative) numeric vector with length \code{nrow(c)}. Each element is the aspect ratio of the GC system of a galaxy.}

\item{n}{A (non-negative) numeric vector with length \code{nrow(c)}. Each element is the Sersic index of the GC system of a galaxy.}

\item{theta}{A numeric vector with length \code{nrow(c)}. Each element is within (0, 2*pi) and is the orientation angle of the GC system of a galaxy.}

\item{mu}{A numeric vector with length \code{nrow(c) + 1}. Each element is the GCLF TO point of each of the \code{nrow(c) + 1} GC sub-populations.}

\item{sigma}{A numeric vector with length \code{nrow(c) + 1}. Each element is the GCLF dispersion of each of the \code{nrow(c) + 1} GC sub-populations.}
}
\value{
A data frame with four columns. \code{x}, \code{y} columns give the locations of the simulated GCs. \code{M} gives the GC magnitudes. \code{id} is the identifier for GC sub-population; \code{id = 0} means GCs from the IGM, while others are from galaxies.
}
\description{
Simulate GC locations and magnitudes (without measurement uncertainty) based on a Sersic profile and Gaussian GCLF.
}
\keyword{Patterns}
\keyword{Point}
\keyword{Sersic}
\keyword{Simulate}
