% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_Y_noise.R
\name{simulate_Y_noisy}
\alias{simulate_Y_noisy}
\title{Simulate GC locations and magnitudes with noisy magnitudes based on a Sersic profile and Gaussian GCLF.}
\usage{
simulate_Y_noisy(
  S,
  l0,
  c,
  N,
  R_eff,
  e,
  n,
  theta,
  mu,
  sigma,
  m0 = 25.5,
  b0 = 0.0884,
  b1 = 0.645
)
}
\arguments{
\item{S}{A \code{SpatialPolygonsDataFrame} that gives the spatial domain on which the point pattern resides.}

\item{l0}{A positive numeric value. The intensity (/kpc^2) of GCs in the IGM.}

\item{c}{A matrix with two columns for the central locations of galaxies in \code{S}. The first column is the x coordinate of the galactic centers, the second column is the y coordinate.}

\item{N}{A (non-negative) numeric vector with length \code{nrow(c)}. Each element is the mean number of GCs in a galaxy.}

\item{R_eff}{A (non-negative) numeric vector with length \code{nrow(c)}. Each element is the half-number radius of the GC system of a galaxy.}

\item{e}{A (non-negative) numeric vector with length \code{nrow(c)}. Each element is the aspect ratio of the GC system of a galaxy.}

\item{n}{A (non-negative) numeric vector with length \code{nrow(c)}. Each element is the Sersic index of the GC system of a galaxy.}

\item{theta}{A numeric vector with length \code{nrow(c)}. Each element is within (0, 2*pi) and is the orientation angle of the GC system of a galaxy.}

\item{mu}{A numeric vector with length \code{nrow(c) + 1}. Each element is the GCLF TO point of each of the \code{nrow(c) + 1} GC sub-populations.}

\item{sigma}{A numeric vector with length \code{nrow(c) + 1}. Each element is the GCLF dispersion of each of the \code{nrow(c) + 1} GC sub-populations.}

\item{m0}{A numeric value. The offset magnitude of the photometric measurement uncertainty.}

\item{b0}{A numeric value. The uncertainty when \code{M = m0}.}

\item{b1}{Rate at which the uncertainty increases with \code{M}.}
}
\value{
A data frame with six columns. \code{x}, \code{y} columns give the locations of the simulated GCs. \code{M} gives the noisy GC magnitudes. \code{Mt} gives the true GC magnitudes. \code{id} is the identifier for GC sub-population; \code{id = 0} means GCs from the IGM, while others are from galaxies. \code{e_M} gives the measurement uncertainty.
}
\description{
Simulate GC locations and magnitudes with noisy magnitudes based on a Sersic profile and Gaussian GCLF.
}
\keyword{Patterns}
\keyword{Point}
\keyword{Sersic}
\keyword{Simulate}
