% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{phi_eM_cpp}
\alias{phi_eM_cpp}
\title{C++ implementation to calculate the value of the noisy GCLF given the observed noisy magnitude \code{M}.}
\usage{
phi_eM_cpp(M, mu, sigma, m0, b0, b1, n = 20L)
}
\arguments{
\item{M}{A numeric vector. The observed GC magnitude}

\item{mu}{Numeric with length 1. True GCLF TO point}

\item{sigma}{Numeric with length 1. True GCLF dispersion}

\item{m0}{Numeric with length 1. The offset magnitude}

\item{b0}{Numeric with length 1. The uncertainty when M = m0}

\item{b1}{Numeric with length 1. Rate at which the uncertainty increases with M}

\item{n}{Integer. Number of integration points used to marginalize out the true magnitude. Default to 20.}
}
\value{
A numeric vector that gives the value of the noisy GCLF given the observed noisy magnitude \code{M}.
}
\description{
Compute the noisy GCLF evaluated at \code{M} based on \deqn{\varphi(m; \mu_{\text{TO}}, \sigma^2) = \int_{-\infty}^\infty\phi(m; m_t, \sigma_M^2(m_t))\phi(m_t; \mu_{\text{TO}}, \sigma^2)dm_t,}
where \eqn{\phi(m; m_t, \sigma_M^2(m_t))} is a Gaussian density that gives the probability density of the noisy magnitude \eqn{m} given the true magnitude \eqn{m_t}. \eqn{\phi(m_t; \mu_{\text{TO}}, \sigma^2)} is the true GCLF with TO point \eqn{\mu_{\text{TO}}} and dispersion \eqn{\sigma}, evaluated at the true magnitude \eqn{m_t}.
}
